+++ argopy/related/__init__.py
@@ -3,6 +3,7 @@ from .ocean_ops_deployments import OceanOPSDeployments
 from .reference_tables import ArgoNVSReferenceTables
 from .argo_documentation import ArgoDocs
 from .doi_snapshot import ArgoDOI
+from .euroargo_api import get_coriolis_profile_id, get_ea_profile_page
 
 
 #
@@ -13,4 +14,8 @@ __all__ = (
     "ArgoNVSReferenceTables",
     "ArgoDocs",
     "ArgoDOI",
+
+    # Functions:
+    "get_coriolis_profile_id",
+    "get_ea_profile_page",
 )
