+++ argopy/tests/test_fetchers_data_argovis.py
@@ -13,7 +13,7 @@ from argopy.errors import (
     CacheFileNotFound,
     FileSystemHasNoCache,
 )
-from argopy.utilities import is_list_of_strings
+from argopy.utils import is_list_of_strings
 from utils import requires_connected_argovis, safe_to_server_errors
