+++ argopy/tests/test_fetchers_data_erddap.py
@@ -1,7 +1,7 @@
 import logging
 
 from argopy import DataFetcher as ArgoDataFetcher
-from argopy.utilities import is_list_of_strings
+from argopy.utils import is_list_of_strings
 
 import pytest
 import xarray as xr
