+++ argopy/tests/test_fetchers_data_gdac.py
@@ -19,7 +19,8 @@ from argopy.errors import (
     FileSystemHasNoCache,
     FtpPathError,
 )
-from argopy.utilities import is_list_of_strings, isconnected
+from argopy.utilities import isconnected
+from argopy.utils import is_list_of_strings
 from utils import requires_gdac
 from mocked_http import mocked_httpserver, mocked_server_address
 from collections import ChainMap
