+++ argopy/tests/test_stores_fsspec.py
@@ -30,9 +30,11 @@ from argopy.errors import (
     DataNotFound,
 )
 from argopy.utilities import (
+    modified_environ,
+)
+from argopy.utils import (
     is_list_of_datasets,
     is_list_of_dicts,
-    modified_environ,
 )
 from utils import requires_connection, requires_connected_argovis
 from mocked_http import mocked_httpserver, mocked_server_address
