+++ argopy/tests/test_stores_index.py
@@ -16,7 +16,7 @@ from argopy.errors import (
     OptionValueError,
     InvalidDatasetStructure,
 )
-from argopy.utilities import (
+from argopy.utils import (
     is_list_of_strings,
 )
 from argopy.stores.argo_index_pd import indexstore_pandas
