+++ argopy/utilities.py
@@ -81,10 +81,6 @@ path2assets = importlib.util.find_spec('argopy.static.assets').submodule_search_
 
 log = logging.getLogger("argopy.utilities")
 
-with open(os.path.join(path2assets, "data_types.json"), "r") as f:
-    DATA_TYPES = json.load(f)
-
-
 
 def clear_cache(fs=None):
     """ Delete argopy cache folder content """
@@ -223,6 +219,7 @@ def load_dict(ptype):
     else:
         raise ValueError("Invalid dictionary name")
 
+
 def mapp_dict(Adictionnary, Avalue):
     if Avalue not in Adictionnary:
         return "Unknown"
@@ -230,202 +227,8 @@ def mapp_dict(Adictionnary, Avalue):
         return Adictionnary[Avalue]
 
 
-def list_available_data_src():
-    """ List all available data sources """
-    sources = {}
-    try:
-        from .data_fetchers import erddap_data as Erddap_Fetchers
-        # Ensure we're loading the erddap data fetcher with the current options:
-        Erddap_Fetchers.api_server_check = Erddap_Fetchers.api_server_check.replace(Erddap_Fetchers.api_server, OPTIONS['erddap'])
-        Erddap_Fetchers.api_server = OPTIONS['erddap']
-
-        sources["erddap"] = Erddap_Fetchers
-    except Exception:
-        warnings.warn(
-            "An error occurred while loading the ERDDAP data fetcher, "
-            "it will not be available !\n%s\n%s"
-            % (sys.exc_info()[0], sys.exc_info()[1])
-        )
-        pass
-
-    try:
-        from .data_fetchers import argovis_data as ArgoVis_Fetchers
-
-        sources["argovis"] = ArgoVis_Fetchers
-    except Exception:
-        warnings.warn(
-            "An error occurred while loading the ArgoVis data fetcher, "
-            "it will not be available !\n%s\n%s"
-            % (sys.exc_info()[0], sys.exc_info()[1])
-        )
-        pass
-
-    try:
-        from .data_fetchers import gdacftp_data as GDAC_Fetchers
-        # Ensure we're loading the gdac data fetcher with the current options:
-        GDAC_Fetchers.api_server_check = OPTIONS['ftp']
-        GDAC_Fetchers.api_server = OPTIONS['ftp']
-
-        sources["gdac"] = GDAC_Fetchers
-    except Exception:
-        warnings.warn(
-            "An error occurred while loading the GDAC data fetcher, "
-            "it will not be available !\n%s\n%s"
-            % (sys.exc_info()[0], sys.exc_info()[1])
-        )
-        pass
-
-    # return dict(sorted(sources.items()))
-    return sources
-
-
-def list_available_index_src():
-    """ List all available index sources """
-    sources = {}
-    try:
-        from .data_fetchers import erddap_index as Erddap_Fetchers
-        # Ensure we're loading the erddap data fetcher with the current options:
-        Erddap_Fetchers.api_server_check = Erddap_Fetchers.api_server_check.replace(Erddap_Fetchers.api_server, OPTIONS['erddap'])
-        Erddap_Fetchers.api_server = OPTIONS['erddap']
-
-        sources["erddap"] = Erddap_Fetchers
-    except Exception:
-        warnings.warn(
-            "An error occurred while loading the ERDDAP index fetcher, "
-            "it will not be available !\n%s\n%s"
-            % (sys.exc_info()[0], sys.exc_info()[1])
-        )
-        pass
-
-    try:
-        from .data_fetchers import gdacftp_index as GDAC_Fetchers
-        # Ensure we're loading the gdac data fetcher with the current options:
-        GDAC_Fetchers.api_server_check = OPTIONS['ftp']
-        GDAC_Fetchers.api_server = OPTIONS['ftp']
-
-        sources["gdac"] = GDAC_Fetchers
-    except Exception:
-        warnings.warn(
-            "An error occurred while loading the GDAC index fetcher, "
-            "it will not be available !\n%s\n%s"
-            % (sys.exc_info()[0], sys.exc_info()[1])
-        )
-        pass
-
-    return sources
-
-
-def list_standard_variables():
-    """ List of variables for standard users """
-    return [
-        "DATA_MODE",
-        "LATITUDE",
-        "LONGITUDE",
-        "POSITION_QC",
-        "DIRECTION",
-        "PLATFORM_NUMBER",
-        "CYCLE_NUMBER",
-        "PRES",
-        "TEMP",
-        "PSAL",
-        "PRES_QC",
-        "TEMP_QC",
-        "PSAL_QC",
-        "PRES_ADJUSTED",
-        "TEMP_ADJUSTED",
-        "PSAL_ADJUSTED",
-        "PRES_ADJUSTED_QC",
-        "TEMP_ADJUSTED_QC",
-        "PSAL_ADJUSTED_QC",
-        "PRES_ADJUSTED_ERROR",
-        "TEMP_ADJUSTED_ERROR",
-        "PSAL_ADJUSTED_ERROR",
-        "PRES_ERROR",  # can be created from PRES_ADJUSTED_ERROR after a filter_data_mode
-        "TEMP_ERROR",
-        "PSAL_ERROR",
-        "JULD",
-        "JULD_QC",
-        "TIME",
-        "TIME_QC",
-        # "CONFIG_MISSION_NUMBER",
-    ]
-
-
-def list_multiprofile_file_variables():
-    """ List of variables in a netcdf multiprofile file.
-
-        This is for files created by GDAC under <DAC>/<WMO>/<WMO>_prof.nc
-    """
-    return [
-        "CONFIG_MISSION_NUMBER",
-        "CYCLE_NUMBER",
-        "DATA_CENTRE",
-        "DATA_MODE",
-        "DATA_STATE_INDICATOR",
-        "DATA_TYPE",
-        "DATE_CREATION",
-        "DATE_UPDATE",
-        "DC_REFERENCE",
-        "DIRECTION",
-        "FIRMWARE_VERSION",
-        "FLOAT_SERIAL_NO",
-        "FORMAT_VERSION",
-        "HANDBOOK_VERSION",
-        "HISTORY_ACTION",
-        "HISTORY_DATE",
-        "HISTORY_INSTITUTION",
-        "HISTORY_PARAMETER",
-        "HISTORY_PREVIOUS_VALUE",
-        "HISTORY_QCTEST",
-        "HISTORY_REFERENCE",
-        "HISTORY_SOFTWARE",
-        "HISTORY_SOFTWARE_RELEASE",
-        "HISTORY_START_PRES",
-        "HISTORY_STEP",
-        "HISTORY_STOP_PRES",
-        "JULD",
-        "JULD_LOCATION",
-        "JULD_QC",
-        "LATITUDE",
-        "LONGITUDE",
-        "PARAMETER",
-        "PI_NAME",
-        "PLATFORM_NUMBER",
-        "PLATFORM_TYPE",
-        "POSITIONING_SYSTEM",
-        "POSITION_QC",
-        "PRES",
-        "PRES_ADJUSTED",
-        "PRES_ADJUSTED_ERROR",
-        "PRES_ADJUSTED_QC",
-        "PRES_QC",
-        "PROFILE_PRES_QC",
-        "PROFILE_PSAL_QC",
-        "PROFILE_TEMP_QC",
-        "PROJECT_NAME",
-        "PSAL",
-        "PSAL_ADJUSTED",
-        "PSAL_ADJUSTED_ERROR",
-        "PSAL_ADJUSTED_QC",
-        "PSAL_QC",
-        "REFERENCE_DATE_TIME",
-        "SCIENTIFIC_CALIB_COEFFICIENT",
-        "SCIENTIFIC_CALIB_COMMENT",
-        "SCIENTIFIC_CALIB_DATE",
-        "SCIENTIFIC_CALIB_EQUATION",
-        "STATION_PARAMETERS",
-        "TEMP",
-        "TEMP_ADJUSTED",
-        "TEMP_ADJUSTED_ERROR",
-        "TEMP_ADJUSTED_QC",
-        "TEMP_QC",
-        "VERTICAL_SAMPLING_SCHEME",
-        "WMO_INST_TYPE",
-    ]
-
-
 def get_sys_info():
-    "Returns system information as a dict"
+    """Returns system information as a dict"""
 
     blob = []
 
@@ -624,94 +427,6 @@ def show_options(file=sys.stdout):  # noqa: C901
     for k, v in opts.items():
         print(f"{k}: {v}", file=file)
 
-
-def check_gdac_path(path, errors='ignore'):  # noqa: C901
-    """ Check if a path has the expected GDAC ftp structure
-
-        Expected GDAC ftp structure::
-
-            .
-            └── dac
-                ├── aoml
-                ├── ...
-                ├── coriolis
-                ├── ...
-                ├── meds
-                └── nmdis
-
-        This check will return True if at least one DAC sub-folder is found under path/dac/<dac_name>
-
-        Examples::
-        >>> check_gdac_path("https://data-argo.ifremer.fr")  # True
-        >>> check_gdac_path("ftp://ftp.ifremer.fr/ifremer/argo") # True
-        >>> check_gdac_path("ftp://usgodae.org/pub/outgoing/argo") # True
-        >>> check_gdac_path("/home/ref-argo/gdac") # True
-        >>> check_gdac_path("https://www.ifremer.fr") # False
-        >>> check_gdac_path("ftp://usgodae.org/pub/outgoing") # False
-
-        Parameters
-        ----------
-        path: str
-            Path name to check, including access protocol
-        errors: str
-            "ignore" or "raise" (or "warn")
-
-        Returns
-        -------
-        checked: boolean
-            True if at least one DAC folder is found under path/dac/<dac_name>
-            False otherwise
-    """
-    # Create a file system for this path
-    if split_protocol(path)[0] is None:
-        fs = fsspec.filesystem('file')
-    elif 'https' in split_protocol(path)[0]:
-        fs = fsspec.filesystem('http')
-    elif 'ftp' in split_protocol(path)[0]:
-        try:
-            host = split_protocol(path)[-1].split('/')[0]
-            fs = fsspec.filesystem('ftp', host=host)
-        except gaierror:
-            if errors == 'raise':
-                raise FtpPathError("Can't get address info (GAIerror) on '%s'" % host)
-            elif errors == "warn":
-                warnings.warn("Can't get address info (GAIerror) on '%s'" % host)
-                return False
-            else:
-                return False
-    else:
-        raise FtpPathError("Unknown protocol for an Argo GDAC host: %s" % split_protocol(path)[0])
-
-    # dacs = [
-    #     "aoml",
-    #     "bodc",
-    #     "coriolis",
-    #     "csio",
-    #     "csiro",
-    #     "incois",
-    #     "jma",
-    #     "kma",
-    #     "kordi",
-    #     "meds",
-    #     "nmdis",
-    # ]
-
-    # Case 1:
-    check1 = (
-        fs.exists(path)
-        and fs.exists(fs.sep.join([path, "dac"]))
-        # and np.any([fs.exists(fs.sep.join([path, "dac", dac])) for dac in dacs])  # Take too much time on http/ftp GDAC server
-    )
-    if check1:
-        return True
-    elif errors == "raise":
-        raise FtpPathError("This path is not GDAC compliant (no `dac` folder with legitimate sub-folder):\n%s" % path)
-
-    elif errors == "warn":
-        warnings.warn("This path is not GDAC compliant:\n%s" % path)
-        return False
-    else:
-        return False
 
 
 def isconnected(host: str = "https://www.ifremer.fr", maxtry: int = 10):
@@ -1372,16 +1087,6 @@ def format_oneline(s, max_width=65):
         return s
 
 
-def to_list(obj):
-    """Make sure that an expected list is indeed a list"""
-    if not isinstance(obj, list):
-        if isinstance(obj, np.ndarray):
-            obj = list(obj)
-        else:
-            obj = [obj]
-    return obj
-
-
 def warnUnless(ok, txt):
     """Function to raise a warning unless condition is True
 
@@ -1843,157 +1548,6 @@ def argo_split_path(this_path):  # noqa C901
     return dict(sorted(output.items()))
 
 
-class DocInherit(object):
-    """Docstring inheriting method descriptor
-
-    The class itself is also used as a decorator
-
-    Usage:
-
-    class Foo(object):
-        def foo(self):
-            "Frobber"
-            pass
-
-    class Bar(Foo):
-        @doc_inherit
-        def foo(self):
-            pass
-
-    Now, Bar.foo.__doc__ == Bar().foo.__doc__ == Foo.foo.__doc__ == "Frobber"
-
-    src: https://code.activestate.com/recipes/576862/
-    """
-
-    def __init__(self, mthd):
-        self.mthd = mthd
-        self.name = mthd.__name__
-
-    def __get__(self, obj, cls):
-        if obj:
-            return self.get_with_inst(obj, cls)
-        else:
-            return self.get_no_inst(cls)
-
-    def get_with_inst(self, obj, cls):
-
-        overridden = getattr(super(cls, obj), self.name, None)
-
-        @wraps(self.mthd, assigned=('__name__', '__module__'))
-        def f(*args, **kwargs):
-            return self.mthd(obj, *args, **kwargs)
-
-        return self.use_parent_doc(f, overridden)
-
-    def get_no_inst(self, cls):
-
-        for parent in cls.__mro__[1:]:
-            overridden = getattr(parent, self.name, None)
-            if overridden:
-                break
-
-        @wraps(self.mthd, assigned=('__name__', '__module__'))
-        def f(*args, **kwargs):
-            return self.mthd(*args, **kwargs)
-
-        return self.use_parent_doc(f, overridden)
-
-    def use_parent_doc(self, func, source):
-        if source is None:
-            raise NameError("Can't find '%s' in parents" % self.name)
-        func.__doc__ = source.__doc__
-        return func
-
-
-doc_inherit = DocInherit
-
-
-def deprecated(reason):
-    """Deprecation warning decorator.
-
-    This is a decorator which can be used to mark functions
-    as deprecated. It will result in a warning being emitted
-    when the function is used.
-
-    Parameters
-    ----------
-    reason: {str, None}
-        Text message to send with deprecation warning
-
-    Examples
-    --------
-    The @deprecated can be used with a 'reason'.
-
-        .. code-block:: python
-
-           @deprecated("please, use another function")
-           def old_function(x, y):
-             pass
-
-    or without:
-
-        .. code-block:: python
-
-           @deprecated
-           def old_function(x, y):
-             pass
-
-    References
-    ----------
-    https://stackoverflow.com/a/40301488
-    """
-    import inspect
-
-    if isinstance(reason, str):
-
-        def decorator(func1):
-
-            if inspect.isclass(func1):
-                fmt1 = "Call to deprecated class {name} ({reason})."
-            else:
-                fmt1 = "Call to deprecated function {name} ({reason})."
-
-            @wraps(func1)
-            def new_func1(*args, **kwargs):
-                warnings.simplefilter('always', DeprecationWarning)
-                warnings.warn(
-                    fmt1.format(name=func1.__name__, reason=reason),
-                    category=DeprecationWarning,
-                    stacklevel=2
-                )
-                warnings.simplefilter('default', DeprecationWarning)
-                return func1(*args, **kwargs)
-
-            return new_func1
-
-        return decorator
-
-    elif inspect.isclass(reason) or inspect.isfunction(reason):
-
-        func2 = reason
-
-        if inspect.isclass(func2):
-            fmt2 = "Call to deprecated class {name}."
-        else:
-            fmt2 = "Call to deprecated function {name}."
-
-        @wraps(func2)
-        def new_func2(*args, **kwargs):
-            warnings.simplefilter('always', DeprecationWarning)
-            warnings.warn(
-                fmt2.format(name=func2.__name__),
-                category=DeprecationWarning,
-                stacklevel=2
-            )
-            warnings.simplefilter('default', DeprecationWarning)
-            return func2(*args, **kwargs)
-
-        return new_func2
-
-    else:
-        raise TypeError(repr(type(reason)))
-
-
 class RegistryItem(ABC):
     """Prototype for possible custom items in a Registry"""
     @property
@@ -2244,456 +1798,6 @@ class Registry(UserList):
     def copy(self):
         """Return a shallow copy of the registry"""
         return self.__copy__()
-
-
-def get_coriolis_profile_id(WMO, CYC=None, **kwargs):
-    """ Return a :class:`pandas.DataFrame` with CORIOLIS ID of WMO/CYC profile pairs
-
-        This method get ID by requesting the dataselection.euro-argo.eu trajectory API.
-
-        Parameters
-        ----------
-        WMO: int, list(int)
-            Define the list of Argo floats. This is a list of integers with WMO float identifiers.
-            WMO is the World Meteorological Organization.
-        CYC: int, list(int)
-            Define the list of cycle numbers to load ID for each Argo floats listed in ``WMO``.
-
-        Returns
-        -------
-        :class:`pandas.DataFrame`
-    """
-    WMO_list = check_wmo(WMO)
-    if CYC is not None:
-        CYC_list = check_cyc(CYC)
-    if 'api_server' in kwargs:
-        api_server = kwargs['api_server']
-    elif OPTIONS['server'] is not None:
-        api_server = OPTIONS['server']
-    else:
-        api_server = "https://dataselection.euro-argo.eu/api"
-    URIs = [api_server + "/trajectory/%i" % wmo for wmo in WMO_list]
-
-    def prec(data, url):
-        # Transform trajectory json to dataframe
-        # See: https://dataselection.euro-argo.eu/swagger-ui.html#!/cycle-controller/getCyclesByPlatformCodeUsingGET
-        WMO = check_wmo(url.split("/")[-1])[0]
-        rows = []
-        for profile in data:
-            keys = [x for x in profile.keys() if x not in ["coordinate"]]
-            meta_row = dict((key, profile[key]) for key in keys)
-            for row in profile["coordinate"]:
-                meta_row[row] = profile["coordinate"][row]
-            meta_row["WMO"] = WMO
-            rows.append(meta_row)
-        return pd.DataFrame(rows)
-
-    from .stores import httpstore
-    fs = httpstore(cache=True, cachedir=OPTIONS['cachedir'])
-    data = fs.open_mfjson(URIs, preprocess=prec, errors="raise", url_follow=True)
-
-    # Merge results (list of dataframe):
-    key_map = {
-        "id": "ID",
-        "lat": "LATITUDE",
-        "lon": "LONGITUDE",
-        "cvNumber": "CYCLE_NUMBER",
-        "level": "level",
-        "WMO": "PLATFORM_NUMBER",
-    }
-    for i, df in enumerate(data):
-        df = df.reset_index()
-        df = df.rename(columns=key_map)
-        df = df[[value for value in key_map.values() if value in df.columns]]
-        data[i] = df
-    df = pd.concat(data, ignore_index=True)
-    df.sort_values(by=["PLATFORM_NUMBER", "CYCLE_NUMBER"], inplace=True)
-    df = df.reset_index(drop=True)
-    # df = df.set_index(["PLATFORM_NUMBER", "CYCLE_NUMBER"])
-    df = df.astype({"ID": int})
-    if CYC is not None:
-        df = pd.concat([df[df["CYCLE_NUMBER"] == cyc] for cyc in CYC_list]).reset_index(
-            drop=True
-        )
-    return df[
-        ["PLATFORM_NUMBER", "CYCLE_NUMBER", "ID", "LATITUDE", "LONGITUDE", "level"]
-    ]
-
-
-def get_ea_profile_page(WMO, CYC=None, **kwargs):
-    """ Return a list of URL
-
-        Parameters
-        ----------
-        WMO: int, list(int)
-            WMO must be an integer or an iterable with elements that can be casted as integers
-        CYC: int, list(int), default (None)
-            CYC must be an integer or an iterable with elements that can be casted as positive integers
-
-        Returns
-        -------
-        list(str)
-
-        See also
-        --------
-        get_coriolis_profile_id
-    """
-    df = get_coriolis_profile_id(WMO, CYC, **kwargs)
-    url = "https://dataselection.euro-argo.eu/cycle/{}"
-    return [url.format(this_id) for this_id in sorted(df["ID"])]
-
-
-@deprecated
-def cast_types(ds):  # noqa: C901
-    """ Make sure variables are of the appropriate types according to Argo
-
-    #todo: This is hard coded, but should be retrieved from an API somewhere.
-    Should be able to handle all possible variables encountered in the Argo dataset.
-
-    Parameter
-    ---------
-    :class:`xarray.DataSet`
-
-    Returns
-    -------
-    :class:`xarray.DataSet`
-    """
-
-    list_str = [
-        "PLATFORM_NUMBER",
-        "DATA_MODE",
-        "DIRECTION",
-        "DATA_CENTRE",
-        "DATA_TYPE",
-        "FORMAT_VERSION",
-        "HANDBOOK_VERSION",
-        "PROJECT_NAME",
-        "PI_NAME",
-        "STATION_PARAMETERS",
-        "DATA_CENTER",
-        "DC_REFERENCE",
-        "DATA_STATE_INDICATOR",
-        "PLATFORM_TYPE",
-        "FIRMWARE_VERSION",
-        "POSITIONING_SYSTEM",
-        "PROFILE_PRES_QC",
-        "PROFILE_PSAL_QC",
-        "PROFILE_TEMP_QC",
-        "PARAMETER",
-        "SCIENTIFIC_CALIB_EQUATION",
-        "SCIENTIFIC_CALIB_COEFFICIENT",
-        "SCIENTIFIC_CALIB_COMMENT",
-        "HISTORY_INSTITUTION",
-        "HISTORY_STEP",
-        "HISTORY_SOFTWARE",
-        "HISTORY_SOFTWARE_RELEASE",
-        "HISTORY_REFERENCE",
-        "HISTORY_QCTEST",
-        "HISTORY_ACTION",
-        "HISTORY_PARAMETER",
-        "VERTICAL_SAMPLING_SCHEME",
-        "FLOAT_SERIAL_NO",
-        "SOURCE",
-        "EXPOCODE",
-        "QCLEVEL",
-    ]
-    list_int = [
-        "PLATFORM_NUMBER",
-        "WMO_INST_TYPE",
-        "WMO_INST_TYPE",
-        "CYCLE_NUMBER",
-        "CONFIG_MISSION_NUMBER",
-    ]
-    list_datetime = [
-        "REFERENCE_DATE_TIME",
-        "DATE_CREATION",
-        "DATE_UPDATE",
-        "JULD",
-        "JULD_LOCATION",
-        "SCIENTIFIC_CALIB_DATE",
-        "HISTORY_DATE",
-        "TIME"
-    ]
-
-    def fix_weird_bytes(x):
-        x = x.replace(b"\xb1", b"+/-")
-        return x
-    fix_weird_bytes = np.vectorize(fix_weird_bytes)
-
-    def cast_this(da, type):
-        """ Low-level casting of DataArray values """
-        try:
-            da.values = da.values.astype(type)
-            da.attrs["casted"] = 1
-        except Exception:
-            msg = "Oops! %s occurred. Fail to cast <%s> into %s for: %s. Encountered unique values: %s" % (sys.exc_info()[0], str(da.dtype), type, da.name, str(np.unique(da)))
-            log.debug(msg)
-        return da
-
-    def cast_this_da(da):
-        """ Cast any DataArray """
-        v = da.name
-        da.attrs["casted"] = 0
-        if v in list_str and da.dtype == "O":  # Object
-            if v in ["SCIENTIFIC_CALIB_COEFFICIENT"]:
-                da.values = fix_weird_bytes(da.values)
-            da = cast_this(da, str)
-
-        if v in list_int:  # and da.dtype == 'O':  # Object
-            da = cast_this(da, np.int32)
-
-        if v in list_datetime and da.dtype == "O":  # Object
-            if (
-                "conventions" in da.attrs
-                and da.attrs["conventions"] == "YYYYMMDDHHMISS"
-            ):
-                if da.size != 0:
-                    if len(da.dims) <= 1:
-                        val = da.astype(str).values.astype("U14")
-                        # This should not happen, but still ! That's real world data
-                        val[val == "              "] = "nan"
-                        da.values = pd.to_datetime(val, format="%Y%m%d%H%M%S")
-                    else:
-                        s = da.stack(dummy_index=da.dims)
-                        val = s.astype(str).values.astype("U14")
-                        # This should not happen, but still ! That's real world data
-                        val[val == ""] = "nan"
-                        val[val == "              "] = "nan"
-                        #
-                        s.values = pd.to_datetime(val, format="%Y%m%d%H%M%S")
-                        da.values = s.unstack("dummy_index")
-                    da = cast_this(da, 'datetime64[s]')
-                else:
-                    da = cast_this(da, 'datetime64[s]')
-
-            elif v == "SCIENTIFIC_CALIB_DATE":
-                da = cast_this(da, str)
-                s = da.stack(dummy_index=da.dims)
-                s.values = pd.to_datetime(s.values, format="%Y%m%d%H%M%S")
-                da.values = (s.unstack("dummy_index")).values
-                da = cast_this(da, 'datetime64[s]')
-
-        if "QC" in v and "PROFILE" not in v and "QCTEST" not in v:
-            if da.dtype == "O":  # convert object to string
-                da = cast_this(da, str)
-
-            # Address weird string values:
-            # (replace missing or nan values by a '0' that will be cast as an integer later
-
-            if da.dtype == "<U3":  # string, len 3 because of a 'nan' somewhere
-                ii = (
-                    da == "   "
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                ii = (
-                    da == "nan"
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                # Get back to regular U1 string
-                da = cast_this(da, np.dtype("U1"))
-
-            if da.dtype == "<U1":  # string
-                ii = (
-                    da == ""
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                ii = (
-                    da == " "
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                ii = (
-                    da == "n"
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-            # finally convert QC strings to integers:
-            da = cast_this(da, np.int32)
-
-        if da.dtype == 'O':
-            # By default, try to cast as float:
-            da = cast_this(da, np.float32)
-
-        if da.dtype != "O":
-            da.attrs["casted"] = 1
-
-        return da
-
-    for v in ds.variables:
-        try:
-            ds[v] = cast_this_da(ds[v])
-        except Exception:
-            print("Oops!", sys.exc_info()[0], "occurred.")
-            print("Fail to cast: %s " % v)
-            print("Encountered unique values:", np.unique(ds[v]))
-            raise
-
-    return ds
-
-
-def cast_Argo_variable_type(ds, overwrite=True):
-    """ Ensure that all dataset variables are of the appropriate types according to Argo references
-
-    Parameter
-    ---------
-    :class:`xarray.DataSet`
-    overwrite: bool, default=True
-        Should we force to re-cast a variable we already casted in this dataset ?
-
-    Returns
-    -------
-    :class:`xarray.DataSet`
-    """
-    def cast_this(da, type):
-        """ Low-level casting of DataArray values """
-        try:
-            da = da.astype(type)
-            # with warnings.catch_warnings():
-            #     warnings.filterwarnings('error')
-            #     try:
-            #         da = da.astype(type)
-            #     except Warning:
-            #         log.debug(type)
-            #         log.debug(da.attrs)
-            da.attrs["casted"] = 1
-        except Exception:
-            print("Oops!", sys.exc_info()[0], "occurred.")
-            print("Fail to cast %s[%s] from '%s' to %s" % (da.name, da.dims, da.dtype, type))
-            try:
-                print("Unique values:", np.unique(da))
-            except Exception:
-                print("Can't read unique values !")
-                pass
-        return da
-
-    def cast_this_da(da, v):
-        """ Cast any Argo DataArray """
-        # print("Casting %s ..." % da.name)
-        da.attrs["casted"] = 0
-
-        if v in DATA_TYPES['data']['str'] and da.dtype == "O":  # Object
-            da = cast_this(da, str)
-
-        if v in DATA_TYPES['data']['int']:  # and da.dtype == 'O':  # Object
-            if "conventions" in da.attrs:
-                convname = "conventions"
-            elif "convention" in da.attrs:
-                convname = "convention"
-            else:
-                convname = None
-            if (
-                    convname in da.attrs
-                    and da.attrs[convname] in ["Argo reference table 19",
-                                               "Argo reference table 21",
-                                               "WMO float identifier : A9IIIII",
-                                               "1...N, 1 : first complete mission",
-                                               ]
-            ):
-                # Some values may be missing, and the _FillValue=" " cannot be casted as an integer.
-                # so, we replace missing values with a 999:
-                val = da.astype(str).values
-                # val[np.where(val == 'nan')] = '999'
-                val[val == 'nan'] = '999'
-                da.values = val
-            da = cast_this(da, float)
-            da = cast_this(da, int)
-
-        if v in DATA_TYPES['data']['datetime'] and da.dtype == "O":  # Object
-            if (
-                    "conventions" in da.attrs
-                    and da.attrs["conventions"] == "YYYYMMDDHHMISS"
-            ):
-                if da.size != 0:
-                    if len(da.dims) <= 1:
-                        val = da.astype(str).values.astype("U14")
-                        # This should not happen, but still ! That's real world data
-                        val[val == "              "] = "nan"
-                        da.values = pd.to_datetime(val, format="%Y%m%d%H%M%S")
-                    else:
-                        s = da.stack(dummy_index=da.dims)
-                        val = s.astype(str).values.astype("U14")
-                        # This should not happen, but still ! That's real world data
-                        val[val == "              "] = "nan"
-                        s.values = pd.to_datetime(val, format="%Y%m%d%H%M%S")
-                        da.values = s.unstack("dummy_index")
-                    da = cast_this(da, 'datetime64[ns]')
-                else:
-                    da = cast_this(da, 'datetime64[ns]')
-
-            elif v == "SCIENTIFIC_CALIB_DATE":
-                da = cast_this(da, str)
-                s = da.stack(dummy_index=da.dims)
-                s.values = pd.to_datetime(s.values, format="%Y%m%d%H%M%S")
-                da.values = (s.unstack("dummy_index")).values
-                da = cast_this(da, 'datetime64[ns]')
-
-        if "QC" in v and "PROFILE" not in v and "QCTEST" not in v:
-            if da.dtype == "O":  # convert object to string
-                da = cast_this(da, str)
-
-            # Address weird string values:
-            # (replace missing or nan values by a '0' that will be cast as an integer later
-            if da.dtype == float:
-                val = da.astype(str).values
-                val[np.where(val == 'nan')] = '0'
-                da.values = val
-                da = cast_this(da, float)
-
-            if da.dtype == "<U3":  # string, len 3 because of a 'nan' somewhere
-                ii = (
-                        da == "   "
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                ii = (
-                        da == "nan"
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                # Get back to regular U1 string
-                da = cast_this(da, np.dtype("U1"))
-
-            if da.dtype == "<U1":  # string
-                ii = (
-                        da == ""
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                ii = (
-                        da == " "
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-                ii = (
-                        da == "n"
-                )  # This should not happen, but still ! That's real world data
-                da = xr.where(ii, "0", da)
-
-            # finally convert QC strings to integers:
-            da = cast_this(da, int)
-
-        if "DATA_MODE" in v:
-            da = cast_this(da, '<U1')
-
-        if da.dtype != "O":
-            da.attrs["casted"] = 1
-
-        return da
-
-    for v in ds.variables:
-        if overwrite or ('casted' in ds[v].attrs and ds[v].attrs['casted'] == 0):
-            try:
-                ds[v] = cast_this_da(ds[v], v)
-            except Exception:
-                print("Oops!", sys.exc_info()[0], "occurred.")
-                print("Fail to cast: %s " % v)
-                print("Encountered unique values:", np.unique(ds[v]))
-                raise
-
-    return ds
 
 
 def log_argopy_callerstack(level='debug'):
