+++ argopy/utils/__init__.py
@@ -5,7 +5,18 @@ from .checkers import (
     is_wmo, check_wmo,
     is_cyc, check_cyc,
     check_index_cols,
+    check_gdac_path,
 )
+from .casting import DATA_TYPES, cast_Argo_variable_type, to_list
+from .decorators import deprecated, doc_inherit
+from .lists import (
+    list_available_data_src,
+    list_available_index_src,
+    list_standard_variables,
+    list_multiprofile_file_variables
+)
+
+
 
 __all__ = (
     # Classes:
@@ -17,4 +28,20 @@ __all__ = (
     "is_wmo", "check_wmo",
     "is_cyc", "check_cyc",
     "check_index_cols",
+    "check_gdac_path",
+
+    # Data type casting:
+    "DATA_TYPES",
+    "cast_Argo_variable_type",
+    "to_list",
+
+    # Decorators:
+    "deprecated",
+    "doc_inherit",
+
+    # Lists:
+    "list_available_data_src",
+    "list_available_index_src",
+    "list_standard_variables",
+    "list_multiprofile_file_variables",
 )
